backend/pyproject.toml
[project]
name = "pricetracker"
version = "0.1.0"
description = "Web scraper for Amazon and Newegg"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi[standard]>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "pyjwt[crypto]>=2.10.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
backend = "fastapi dev src.main:pricetracker"

